Fix build of Jupyter extension and test that it works
If you look in the build logs of the current 0.6.1 package, there's an error
in the `pip install` command because `npm` isn't available. Somewhat
annoyingly, this error doesn't actually cause the installation to exit with an
error.

However, this PR doesn't bump the build number because the 0.6.1 release
tarball doesn't contain the files needed to actually build the JavaScript
files, so we still can't install the extension successfully. We'll need a new
ipyevents tarball for that to work.